# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance Campaign Popup Service with Redis \u0026 PostgreSQL.",
        "title": "Campaign Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/campaigns": {
            "get": {
                "description": "Fetches all campaigns from PostgreSQL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List All Campaigns",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/campaign.Campaign"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a campaign in DB and syncs to Redis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update Campaign",
                "parameters": [
                    {
                        "description": "Campaign Data",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a campaign in DB and syncs to Redis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create New Campaign",
                "parameters": [
                    {
                        "description": "Campaign Data",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a campaign from DB and Redis.",
                "tags": [
                    "Admin"
                ],
                "summary": "Delete Campaign",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/admin/campaigns/detail": {
            "get": {
                "description": "Fetches a single campaign from PostgreSQL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get Campaign Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    }
                }
            }
        },
        "/debug/sync": {
            "post": {
                "description": "Manually triggers synchronization of all campaigns from DB to Redis.",
                "tags": [
                    "Debug"
                ],
                "summary": "Sync DB to Redis",
                "responses": {
                    "200": {
                        "description": "Synced"
                    }
                }
            }
        },
        "/v1/campaigns/impression": {
            "post": {
                "description": "Records that a user has seen a campaign.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Track Impression",
                "parameters": [
                    {
                        "description": "Impression Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ImpressionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/campaigns/popup": {
            "get": {
                "description": "Determines the best campaign popup for a user based on priority, time, and targeting.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Get Popup for User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/campaign.Campaign"
                        }
                    },
                    "204": {
                        "description": "No Content (No suitable campaign)"
                    },
                    "400": {
                        "description": "Invalid User ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "campaign.Campaign": {
            "type": "object",
            "properties": {
                "action_url": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "description": "For DB/Admin",
                    "type": "boolean"
                },
                "max_frequency": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "target_segment": {
                    "description": "If TargetType == SEGMENT",
                    "type": "string"
                },
                "target_type": {
                    "$ref": "#/definitions/campaign.TargetType"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "campaign.TargetType": {
            "type": "string",
            "enum": [
                "ALL",
                "SEGMENT"
            ],
            "x-enum-varnames": [
                "TargetTypeAll",
                "TargetTypeSegment"
            ]
        },
        "main.ImpressionRequest": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}